Install mkdocs deps to user
Uses the `--user` flag to `pip install` to ensure pip installs local to the
user, so that permissions will be correct. Also adds `$HOME/.local/bin` to the
path, so that `mkdocs` can be executed once installed.

Finally, caches the `$HOME/.local` directory, so that subsequent builds do not
need to install the deps again.